back/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-restframework==0.0.1
sqlparse==0.5.3
tzdata==2025.2